+++ python/tests/t_cluster_tag.py
@@ -84,19 +84,19 @@ spec:
         clusters = result.json["clusters"]
         
         cluster_1 = clusters["cluster_clustertagtest_http_target1_default"]
-        self.assert_cluster(cluster_1, "clustertagtest-http-target1.default")
+        self.assert_cluster(cluster_1, "clustertagtest-http-target1")
 
         cluster_2 = clusters["cluster_tag_1_clustertagtest_http_target1_default"]
-        self.assert_cluster(cluster_2, "clustertagtest-http-target1.default")
+        self.assert_cluster(cluster_2, "clustertagtest-http-target1")
 
         cluster_3 = clusters["cluster_tag_2_clustertagtest_http_target1_default"]
-        self.assert_cluster(cluster_3, "clustertagtest-http-target1.default")
+        self.assert_cluster(cluster_3, "clustertagtest-http-target1")
 
         cluster_4 = clusters["cluster_tag_2_clustertagtest_http_target2_default"]
-        self.assert_cluster(cluster_4, "clustertagtest-http-target2.default")
+        self.assert_cluster(cluster_4, "clustertagtest-http-target2")
 
         cluster_5 = clusters["cluster_some_really_long_tag_that_is_rea-0"]
-        self.assert_cluster(cluster_5, "clustertagtest-http-target1.default")
+        self.assert_cluster(cluster_5, "clustertagtest-http-target1")
 
         cluster_6 = clusters["cluster_some_really_long_tag_that_is_rea-1"]
-        self.assert_cluster(cluster_6, "clustertagtest-http-target2.default")
+        self.assert_cluster(cluster_6, "clustertagtest-http-target2")